% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapref_STAN.R
\name{adapref_sampling}
\alias{adapref_sampling}
\title{Adaptive preferential sampling method_HMC using stan}
\usage{
adapref_sampling(
  prior.c = "normal",
  prior.s = "horseshoe",
  order.c = 1,
  order.s = 1,
  zeta.c = 0.01,
  zeta.s = 0.01,
  fit = NA,
  data,
  pars = NA,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  control = list(adapt_delta = 0.95, max_treedepth = 12),
  save.loglik = FALSE,
  ...
)
}
\arguments{
\item{prior.c}{horseshoe or normal (gaussian) field on effective pop size}

\item{prior.s}{horseshoe or normal (gaussian) field on beta}

\item{order.s}{order field on beta}

\item{data}{list containing object or list containing vectors of coalescent
times \code{coal_times}, sampling times \code{samp_times},  number
sampled per sampling time \code{n_sampled}, \code{grid} grid points.\#' @param zeta1 precision parameter field on effective pop size}

\item{iter}{number of interations}

\item{warmup}{burn in}

\item{thin}{thinning}

\item{control}{list control variables, typicall list(adapt_delta=0.98, max_treedepth=12)}

\item{save.loglik}{save log like if one wants to compute  = T}

\item{zeta1}{precision parameter field on effective pop size}

\item{zeta2}{precision parameter field on beta}

\item{par}{vector parameter to save}

\item{chain}{number of chains to run}

\item{derivative}{logical whether to calculate estimates of the
log-derivative.}

\item{forward}{logical whether to use the finite difference approximations of
the log-derivative as a forward or backward derivative.}
}
\value{
Effective pop size and beta posterior- Stan output, to be passed into extract_out
in order to have format for plotting
}
\description{
Adaptive preferential sampling method_HMC using stan
}
