% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapref_INLA.R, R/adapref_INLA_phylodyn.R
\name{BNPR_PS_ada}
\alias{BNPR_PS_ada}
\title{Adaptive preferential sampling method_INLA approximation}
\usage{
BNPR_PS_ada(
  data,
  lengthout = 100,
  zeta1 = 0.01,
  zeta2 = 0.01,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE
)

BNPR_PS_ada(
  data,
  lengthout = 100,
  zeta1 = 0.01,
  zeta2 = 0.01,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE
)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent
times \code{coal_times}, sampling times \code{samp_times}, and number
sampled per sampling time \code{n_sampled}.}

\item{lengthout}{number of grid points.}

\item{zeta1}{precision parameter field on effective pop size}

\item{zeta2}{precision parameter field on beta}

\item{simplify}{logical whether to fully bucket all Poisson points.}

\item{derivative}{logical whether to calculate estimates of the
log-derivative.}

\item{forward}{logical whether to use the finite difference approximations of
the log-derivative as a forward or backward derivative.}
}
\value{
Effective pop size and beta posterior,
\code{result} contains the INLA output, \code{data} contains the
information passed to INLA, \code{grid}  grid boundaries,
\code{x}  grid mid points \code{effpop} median effective population size estimates,
\code{effpop025} and \code{effpop975}  2.5th and 97.5th
posterior percentiles, \code{summary} contains a data.frame of the
estimates, and \code{derivative} (if \code{derivative = TRUE})  log-derivative,
\code{sampInt} median beta, \code{sampIntmean} mean beta,
\code{sampInt975} and \code{sampInt025}97.5th and 2.5th posterior percentile beta

Effective pop size and beta posterior,
\code{result} contains the INLA output, \code{data} contains the
information passed to INLA, \code{grid}  grid boundaries,
\code{x}  grid mid points \code{effpop} median effective population size estimates,
\code{effpop025} and \code{effpop975}  2.5th and 97.5th
posterior percentiles, \code{summary} contains a data.frame of the
estimates, and \code{derivative} (if \code{derivative = TRUE})  log-derivative,
\code{sampInt} median beta, \code{sampIntmean} mean beta,
\code{sampInt975} and \code{sampInt025}97.5th and 2.5th posterior percentile beta
}
\description{
Adaptive preferential sampling method_INLA approximation

Adaptive preferential sampling method_INLA approximation
}
