% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapref_STAN.R
\name{samp_and_coalescent_data}
\alias{samp_and_coalescent_data}
\title{Prepare the format neeed for HMC}
\usage{
samp_and_coalescent_data(samp_times, n_sampled, coal_times, grid)
}
\arguments{
\item{grid}{}
}
\value{
all the information needed into the adapref_sampling function
}
\description{
Prepare the format neeed for HMC
}
