% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sampsim_thin}
\alias{sampsim_thin}
\title{Sample sampling times from a Poisson process}
\usage{
sampsim_thin(max.n, traj, xlim, ...)
}
\arguments{
\item{max.n}{number of samples requires}

\item{traj}{function describe the rate of Poisson proces}

\item{xlim}{time interval in which to do sampling}

\item{...}{additional arguments to be passed to the traj function}
}
\description{
Sample sampling times from a Poisson process
}
