% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_prec_parameters}
\alias{set_prec_parameters}
\title{Precision parameters empirica proposals for the two fields}
\usage{
set_prec_parameters(
  phylo,
  ncell,
  upBound = NULL,
  alpha = 0.05,
  order.c = 1,
  order.s = 1
)
}
\arguments{
\item{phylo}{\code{phylo} object or list containing vectors of coalescent
times \code{coal_times}, sampling times \code{samp_times}, and number
sampled per sampling time \code{n_sampled}.}

\item{ncell}{number of grid points.}

\item{alpha}{quantile to consider}

\item{order.c}{order field on effective population size}

\item{order.s}{order field on beta}
}
\value{
Precision parameters,
\code{zeta1}  and \code{zeta2}
}
\description{
Precision parameters empirica proposals for the two fields
}
