% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapref_STAN.R
\name{samp_and_coalescent_data}
\alias{samp_and_coalescent_data}
\title{Prepare the format neeed for HMC}
\usage{
samp_and_coalescent_data(samp_times, n_sampled, coal_times, grid)
}
\arguments{
\item{samp_times}{vector of sampling times}

\item{n_sampled}{vector of number of samples per sampling time}

\item{coal_times}{vector of coalescent time}

\item{grid}{grid points}
}
\value{
all the information needed into the adapref_sampling function
}
\description{
Prepare the format neeed for HMC
}
